prove10 - input validations on server

+++ server.js
@@ -31,6 +31,40 @@ app.listen(app.get('port'), function() {
 	console.log('Node app is running on port', app.get('port'));
 });
 
+//https://gist.github.com/rutcreate/03ff3f9bd5f414465322
+function isInt(n){
+	return Number(n).toString() === n.toString() && n % 1 === 0;
+}
+
+//https://gist.github.com/rutcreate/03ff3f9bd5f414465322
+function isFloat(n){
+	return Number(n).toString() === n.toString() && n % 1 !== 0;
+}
+
+function isValidDate(d){
+	var year = d.getYear();
+	var month = d.getMonth();
+	var day = d.getDate();
+	const date = new Date(`${year}-${month}-${day}`);
+	return (Boolean(+date) && date.getDate().toString() === day);
+}
+
+function isValidType(t){
+	return ((isInt(t) == true) AND (t >= 0) AND (t <= 7) AND (t != null) AND (t != "undefined"));
+}
+
+function isValidName(n){
+	return ((n != null) AND (n != "undefined"));
+}
+
+function isValidAmount(a){
+	return ((isFloat(a) == true) AND (a >= 0) AND (a != null) AND (a != "undefined"));
+}
+
+function isValidID(i){
+	return ((isInt(i) == true) AND (i >= 0) AND (i != null) AND (i != "undefined"));
+}
+
 /* ----- addMovement-----------------------------------
 Data from client: type, date, name, amount
 Data returned: ExpenseID or null
@@ -41,19 +75,24 @@ function addMovement(req, res) {
 	var date = query.date;
 	var name = query.name;
 	var amount = query.amount;
-	console.log("Adding movement: " + name);
-	var sql = "INSERT INTO movements(typeID, name, movementDate, amount) VALUES ($1, $2, $3, $4)";
-	var params = [type, name, date, amount];
+	if (isValidType(type) AND isValidDate(date) AND isValidName(name)) {
+		console.log("Adding movement: " + name);
+		var sql = "INSERT INTO movements(typeID, name, movementDate, amount) VALUES ($1, $2, $3, $4)";
+		var params = [type, name, date, amount];
 
-	dbTransaction(sql, params, function(error, result) {
-		if (error || result == null) {
-			res.writeHead(500);
-		} else {
-			res.writeHead(200, {'Content-Type': 'application/json'});
-			console.log(JSON.stringify(result));
-            res.end();
-		}
-	})
+		dbTransaction(sql, params, function(error, result) {
+			if (error || result == null) {
+				res.writeHead(500);
+			} else {
+				res.writeHead(200, {'Content-Type': 'application/json'});
+				res.end(JSON.stringify(result));
+			}
+		})
+	}
+	else
+	{
+		res.status(500).json({success: false, data: "Invalid input."});
+	}
 }
 /* ----- getMovement----------------------------------- 
 Data from client: Month, Type
@@ -61,23 +100,27 @@ Data returned: List of Expenses for the given type
 -------------------------------------------------------*/
 function getMovement(req, res) {
 	var query = req.query;
-	console.log(query);
 	var d = new Date(2018, query.month, 0);
-    var enddate = '2018/'+ query.month +'/' + d.getDate();
+	var enddate = '2018/'+ query.month +'/' + d.getDate();
 	var startdate = '2018/' + query.month + '/01';
 	var type = query.type;
-	console.log("getting movements of " + type + " from " + d.getMonth());
+	if (isValidDate(startdate) AND isValidDate(enddate) AND isValidType(type)) {
+		console.log("getting movements of type" + type + " from month" + d.getMonth());
+		var sql = "SELECT typeID, name, movementDate, amount FROM movements WHERE typeID = $1 AND movementDate BETWEEN $2 AND $3";
+		var params = [type, startdate, enddate];
 
-	var sql = "SELECT typeID, name, movementDate, amount FROM movements WHERE typeID = $1 AND movementDate BETWEEN $2 AND $3";
-	var params = [type, startdate, enddate];
-
-	dbTransaction(sql, params, function(error, result) {
-		 if (error || result == null) {
-		 	 res.status(500).json({success: false, data: error});
-		  } else {
-			 res.status(200).json(result);
-		 }
-	})
+		dbTransaction(sql, params, function(error, result) {
+			if (error || result == null) {
+				res.status(500).json({success: false, data: error});
+			} else {
+				res.status(200).json(result);
+			}
+		})
+	}
+	else
+	{
+		res.status(500).json({success: false, data: "Invalid input."});
+	}
 }
 /* ----- getAvg----------------------------------------
 Data from client: Type
@@ -85,18 +128,23 @@ Data returned: Avg expense on a given category
 -------------------------------------------------------*/
 function getAvg(req, res) {
 	var type = req.query.type;
-	console.log("Getting avg from type: " + type);
+	if (isValidType(type)) {
+		console.log("Getting avg from type: " + type);
+		var sql = "SELECT AVG(amount) FROM movements WHERE typeID = $1";
+		var params = [type];
 
-	var sql = "SELECT AVG(amount) FROM movements WHERE typeID = $1";
-	var params = [type];
-
-	dbTransaction(sql, params, function(error, result) {
-		 if (error || result == null) {
-		 	 res.status(500).json({success: false, data: error});
-		  } else {
-			 res.status(200).json(result);
-		 }
-	})
+		dbTransaction(sql, params, function(error, result) {
+			if (error || result == null) {
+				res.status(500).json({success: false, data: error});
+			} else {
+				res.status(200).json(result);
+			}
+		})
+	}
+	else
+	{
+		res.status(500).json({success: false, data: "Invalid input."});
+	}
 }
 /* ----- getTotal--------------------------------------
 Data from client: Type, Month
@@ -105,22 +153,27 @@ Data returned: Total expenses on a given category
 function getTotal(req, res) {
 	var query = req.query;
 	var d = new Date(2018, query.month, 0);
-    var enddate = '2018/'+ query.month +'/' + d.getDate();
+	var enddate = '2018/'+ query.month +'/' + d.getDate();
 	var startdate = '2018/' + query.month + '/01';
 	var type = query.type;
-	var total = 100.6;
-	console.log("getting total from type: "+ type);
 
-	var sql = "SELECT SUM(amount) FROM movements WHERE typeID = $1 AND movementDate BETWEEN $2 AND $3";
-	var params = [type, startdate, enddate];
+	if (isValidDate(startdate) AND isValidDate(enddate) AND isValidType(type)) {
+		console.log("getting sum of movements from type: " + type);
+		var sql = "SELECT SUM(amount) FROM movements WHERE typeID = $1 AND movementDate BETWEEN $2 AND $3";
+		var params = [type, startdate, enddate];
 
-	dbTransaction(sql, params, function(error, result) {
-		 if (error || result == null) {
-		 	 res.status(500).json({success: false, data: error});
-		  } else {
-			 res.status(200).json(result);
-		 }
-	})
+		dbTransaction(sql, params, function(error, result) {
+			if (error || result == null) {
+				res.status(500).json({success: false, data: error});
+			} else {
+				res.status(200).json(result);
+			}
+		})
+	}
+	else
+	{
+		res.status(500).json({success: false, data: "Invalid input."});
+	}
 }
 /* ----- modifyMovement---------------------------------
 Data from client: MovementID, type, name, amount
@@ -133,36 +186,47 @@ function modifyMovement(req, res) {
 	var name = query.name;
 	var date = query.date;
 	var amount = query.amount;
-	console.log("Modifying movement: "+ name);
-	var sql = "UPDATE movements SET typeID = $1, name = $2, movementDate = $3, amount = $4 WHERE movementID = $5";
-	var params = [type, name, date, amount, movementID];
+	if (isValidID(movementID) AND isValidType(type) AND isValidName(name) AND isValidDate(date) AND isValidAmount(amount)) {
+		console.log("Modifying movement: "+ name);
+		var sql = "UPDATE movements SET typeID = $1, name = $2, movementDate = $3, amount = $4 WHERE movementID = $5";
+		var params = [type, name, date, amount, movementID];
 
-	dbTransaction(sql, params, function(error, result) {
-		 if (error || result == null) {
-		 	 res.status(500).json({success: false, data: error});
-		  } else {
-			 res.status(200).json(result);
-		 }
-	})
+		dbTransaction(sql, params, function(error, result) {
+			if (error || result == null) {
+				res.status(500).json({success: false, data: error});
+			} else {
+				res.status(200).json(result);
+			}
+		})
 	}
+	else
+	{
+		res.status(500).json({success: false, data: "Invalid input."});
+	}
+}
 /* ----- deleteMovement--------------------------------
 Data from client: MovementID
 Data returned: true/false depending on successful deletion
 -------------------------------------------------------*/
 function deleteMovement(req, res) {
 	var movementID = req.query.movementID;
-	console.log("Deleting movement: " + movementID);
-	var sql = "DELETE FROM movements WHERE movementID = $1";
-	var params = [movementID];
-
-	dbTransaction(sql, params, function(error, result) {
-		 if (error || result == null) {
-		 	 res.status(500).json({success: false, data: error});
-		  } else {
-			 res.status(200).json(result);
-		 }
-	})
+	if (isValidID(movementID)) {
+		console.log("Deleting movement: " + movementID);
+		var sql = "DELETE FROM movements WHERE movementID = $1";
+		var params = [movementID];
+		dbTransaction(sql, params, function(error, result) {
+			if (error || result == null) {
+				res.status(500).json({success: false, data: error});
+			} else {
+				res.status(200).json(result);
+			}
+		})
 	}
+	else
+	{
+		res.status(500).json({success: false, data: "Invalid input."});
+	}
+}
 
 function dbTransaction(sql, params, callback) {
 	pool.query(sql, params, function(err, result) {
@@ -170,9 +234,7 @@ function dbTransaction(sql, params, callback) {
 			console.log("Transaction failed: " + err);
 			callback(err, null);
 		}
-
-		console.log("Transaction successful: " + JSON.stringify(result.rows));
-
+		console.log("Transaction successful: " + result);
 		callback(null, result.rows);
 	});
